+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - updated project dir normalization for comparing to avoid infinite projects
 - added support for setting CORS allowed origins
 - added obsidian theme made by GLM-5.1's taste
+- updated generate text function to properly use cache control
 
 ## [0.62.0]